more refactoring of mpl-methods into BrowserBase
